requirements.txt
openai==1.72.0
xinference==1.4.1
numpy==1.26.4
faiss-cpu==1.8.0
tqdm==4.67.1
jsonlines==4.0.0
gradio==5.23.1
packaging==24.2
transformers==4.51.1
modelscope==1.24.1
torch==2.5.1+cu124
psutil==7.0.0
matplotlib==3.10.1
huggingface-hub==0.30.2
requests==2.32.3
datasets==3.2.0
uvicorn==0.34.0
aiohttp==3.11.16
fastapi==0.115.12
pillow==11.1.0
dacite==1.9.2
accelerate==1.6.0
peft==0.15.1
pandas==2.2.3
tensorboard==2.19.0
safetensors==0.5.3
einops==0.8.1
trl==0.16.1
setuptools==69.5.1